refactor: split up unit and e2e tests
